+++ xformers/profiler/profiler.py
@@ -17,15 +17,8 @@ import torch.cuda.memory
 import torch.cuda.nvtx
 import torch.nn as nn
 import torch.profiler
-import torch.utils.hooks
 
 logger = logging.getLogger(__name__)
-
-
-def _normalize_tuple(x):
-    if not isinstance(x, tuple):
-        return (x,)
-    return x
 
 
 class NsightProfiler:
@@ -41,12 +34,10 @@ class NsightProfiler:
         # TODO figure out if there is a way to know if nsys is launched at this point
 
     def __enter__(self):
-        self.main_profiler._install_hooks()
         torch.cuda.profiler.start()
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         torch.cuda.profiler.stop()
-        self.main_profiler._remove_hooks()
 
     def step(self) -> None:
         pass
@@ -174,6 +165,7 @@ class _Profiler:
         module: Optional[nn.Module],
     ) -> None:
         self.check_schedule(schedule)
+        self.schedule = schedule
         self.done_steps = 0
         self.output_dir = Path(output_dir).absolute()
         self.output_dir.mkdir(exist_ok=True, parents=True)
@@ -182,9 +174,6 @@ class _Profiler:
             self.worker_name = "{}_{}".format(socket.gethostname(), str(os.getpid()))
 
         self.module = weakref.ref(module if module is not None else nn.Module())
-        self.parents = ["Global"]
-        self.hooks: List[torch.utils.hooks.RemovableHandle] = []
-        self.hooks_refcount = 0
         self.profilers: List[_ProfilerState] = sorted(
             [_ProfilerState(cls, begin, end) for cls, begin, end in schedule],
             key=lambda x: x.iter_begin,
@@ -250,79 +239,6 @@ class _Profiler:
             return folder / f"{self.done_steps:06}_{self.worker_name}{file.suffix}"
         return self.output_dir / f"{self.done_steps:06}_{filename}"
 
-    def _install_hooks(self) -> None:
-        self.hooks_refcount += 1
-        # Already installed
-        if self.hooks:
-            return
-        module = self.module()
-        if module is None:
-            return
-        for name, sub_mod in module.named_modules():
-            if name == "":
-                continue
-            name = name.split(".")[-1]
-            self.hooks += [
-                sub_mod.register_forward_pre_hook(self._enter_module_hook(name)),
-                sub_mod.register_forward_hook(self._exit_module_hook(name)),
-            ]
-
-    def _remove_hooks(self) -> None:
-        self.hooks_refcount -= 1
-        if self.hooks_refcount == 0:
-            for h in self.hooks:
-                h.remove()
-
-    def _enter_module_hook(self, name):
-        class PopState(torch.autograd.Function):
-            @staticmethod
-            def forward(ctx, *args):
-                if len(args) == 1:
-                    return args[0]
-                return args
-
-            @staticmethod
-            def backward(ctx, *grad_outs):
-                self._exit_module(name)
-                return grad_outs
-
-        def f(module, inputs):
-            self._enter_module(name)
-            inputs = _normalize_tuple(inputs)
-            out = PopState.apply(*inputs)
-            return out
-
-        return f
-
-    def _exit_module_hook(self, name):
-        class PushState(torch.autograd.Function):
-            @staticmethod
-            def forward(ctx, *args):
-                if len(args) == 1:
-                    return args[0]
-                return args
-
-            @staticmethod
-            def backward(ctx, *grad_outs):
-                self._enter_module(name)
-                return grad_outs
-
-        def f(module, inputs, outputs):
-            self._exit_module(name)
-            outputs = _normalize_tuple(outputs)
-            return PushState.apply(*outputs)
-
-        return f
-
-    def _enter_module(self, name) -> None:
-        self.parents.append(name)
-        torch.cuda.nvtx.range_push(name)
-
-    def _exit_module(self, name) -> None:
-        torch.cuda.nvtx.range_pop()
-        assert self.parents[-1] == name
-        self.parents.pop()
-
     def start(self):
         self.__enter__()
 
@@ -349,7 +265,6 @@ class _Profiler:
         self.done_steps += 1
 
         if self.done_steps <= self.last_step:
-            self.parents = ["Global"]
             self.update_profilers_on_step()
         if self.done_steps == self.last_step:
             logger.info("xFormers profiler done. %s", self.format_summary())